+++ components/notification/api/src/main/java/com/flipperdevices/notification/model/UpdateNotificationState.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.notification.model
+
+enum class UpdateNotificationState {
+    ENABLED,
+    DISABLED,
+    IN_PROGRESS
+}