+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/VersionCategory.kt
@@ -24,7 +24,7 @@ import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.settings.impl.R
-import com.flipperdevices.settings.impl.composable.elements.SimpleElement
+import com.flipperdevices.settings.impl.composable.components.SimpleElement
 import com.flipperdevices.settings.impl.viewmodels.VersionViewModel
 import tangle.viewmodel.compose.tangleViewModel