+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/components/UrlElement.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.settings.impl.composable.elements
+package com.flipperdevices.settings.impl.composable.components
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes