+++ instances/app/src/main/java/com/flipperdevices/app/FlipperApplication.kt
@@ -10,12 +10,20 @@ import com.flipperdevices.core.activityholder.CurrentActivityHolder
 import com.flipperdevices.core.di.ApplicationParams
 import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.di.provideDelegate
+import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
 import com.flipperdevices.singleactivity.impl.SingleActivity
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
 import tangle.inject.TangleGraph
 import timber.log.Timber
 
-class FlipperApplication : Application(), ImageLoaderFactory {
+class FlipperApplication : Application(), ImageLoaderFactory, LogTagProvider {
+    override val TAG = "FlipperApplication"
+
+    private val applicationScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
     override fun onCreate() {
         super.onCreate()
 
@@ -25,6 +33,7 @@ class FlipperApplication : Application(), ImageLoaderFactory {
             .create(
                 context = this,
                 application = this,
+                scope = applicationScope,
                 ApplicationParams(
                     startApplicationClass = SingleActivity::class,
                     version = BuildConfig.VERSION_NAME
@@ -48,6 +57,13 @@ class FlipperApplication : Application(), ImageLoaderFactory {
         val component = ComponentHolder.component<MainComponent>()
         val synchronizationApi by component.synchronizationApi
         synchronizationApi.startSynchronization()
+        try {
+            val notificationApi by component.notificationApi
+            notificationApi.init()
+        } catch (e: Exception) {
+            error(e) { "Failed init notification api" }
+        }
+        component.permissionRequestHandlerImpl.get().register(this)
     }
 
     override fun newImageLoader(): ImageLoader {